+++ src/main/kotlin/picklab/backend/auth/entrypoint/AuthApi.kt
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.annotation.AuthenticationPrincipal
-import picklab.backend.auth.domain.MemberPrincipal
+import picklab.backend.common.model.MemberPrincipal
 import picklab.backend.common.model.ResponseWrapper
 import picklab.backend.member.domain.enums.SocialType
