+++ src/main/kotlin/picklab/backend/auth/entrypoint/AuthController.kt
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import picklab.backend.auth.application.AuthUseCase
 import picklab.backend.auth.application.OAuthProviderResolver
-import picklab.backend.auth.domain.MemberPrincipal
 import picklab.backend.auth.infrastructure.AuthCookieCreator
+import picklab.backend.common.model.MemberPrincipal
 import picklab.backend.common.model.ResponseWrapper
 import picklab.backend.common.model.SuccessCode
 import picklab.backend.member.domain.enums.SocialType
