+++ src/main/kotlin/picklab/backend/common/model/MemberPrincipal.kt
@@ -1,4 +1,4 @@
-package picklab.backend.auth.domain
+package picklab.backend.common.model
 
 data class MemberPrincipal(
     val memberId: Long,
